core/components/company/view.py
from rest_framework.generics import ListAPIView, RetrieveAPIView
from core.models import Company
from core.serializers import CompanySerializer


class CompanyListView(ListAPIView):
    queryset = Company.objects.all()
    serializer_class = CompanySerializer

class CompanyDetailView(RetrieveAPIView):
    queryset = Company.objects.all()
    serializer_class = CompanySerializer
